Remove unused SerializableTransient annotation
