+++ modules/commons/src/main/java/org/apache/ignite/internal/util/typedef/internal/LT.java
